pyproject.toml
[project]
name = "skills-orchestrator"
version = "1.1.0"
description = "编译时 Skill 治理工具 — 生成 AGENTS.md，强制执行冲突检测，MCP Server 按需加载"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
keywords = ["claude", "agents", "skills", "mcp", "ai", "llm", "coding-assistant"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Utilities",
]
dependencies = [
    "pyyaml>=6.0",
    "click>=8.0",
    "mcp>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-benchmark>=4.0",
]

[project.urls]
Homepage = "https://github.com/yourname/skills-orchestrator"
Repository = "https://github.com/yourname/skills-orchestrator"
"Bug Tracker" = "https://github.com/yourname/skills-orchestrator/issues"

[project.scripts]
skills-orchestrator = "src.main:cli"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["src", "src.*"]

[tool.setuptools.package-data]
"src" = ["py.typed"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.ruff]
line-length = 100
target-version = "py310"
